+++ CHANGES.rst
@@ -31,6 +31,8 @@ Unreleased
 -   When generating a command's name from a decorated function's name, the
     suffixes ``_command``, ``_cmd``, ``_group``, and ``_grp`` are removed.
     :issue:`2322`
+-   Ensure stderr is completely captured through CliRunner when writing using
+    non-click functions like print
 
 
 Version 8.1.7